Added heap/handles window.
git-svn-id: http://svn.tilp.info/repos/tiemu@720 4f589446-7ac1-0310-82c7-a055b08196ef

+++ tiemu/trunk/src/core/ti_sw/handles.h
@@ -29,6 +29,6 @@
 
 void heap_get_base_address(uint32_t *base);
 void heap_get_size(uint16_t *size);
-void heap_get_block_size(uint8_t handle, uint32_t *addr, uint16_t *size);
+void heap_get_block_size(int handle, uint32_t *addr, uint16_t *size);
 
 #endif
